fixes program so that it runs correctly when built from maven.
This addresses the issue where the generated rounds would not be
shown at all.  The problem was the resources weren't being included
in the built jar.  This slightly restructures the project leverage
the maven standard directory formats that both maven and gradle
default too.  This also allows better integration with maven
and gradle tooling that expect this directory structure.

Now when mvn clean install is run, the generated jar can be run as
normal.